Add stateless (no model) versions for each view
Add extensions for each of the four view base classes, which do not
require a model (so a view is only managed by a controller, but has no
model). These stateless views are meant for scenarios where a view does
not represent any mutable state (e.g. an about window), making the
addition or implementation of a model unnecessary.